feat: add vault reindex helper
