+++ src/search/indexer.ts
@@ -1,0 +1,43 @@
+import fs from 'node:fs/promises';
+import path from 'node:path';
+import matter from 'gray-matter';
+import { CONFIG } from '../config.js';
+import { index, toSearchDoc } from './meili.js';
+import { isMarkdown } from '../utils/paths.js';
+
+async function walk(dir: string): Promise<string[]> {
+    const entries = await fs.readdir(dir, { withFileTypes: true });
+    const files: string[] = [];
+    for (const e of entries) {
+        if (e.name === '.stfolder') continue;
+        if (e.name.startsWith('.')) continue;
+        const abs = path.join(dir, e.name);
+        if (e.isDirectory()) {
+            files.push(...await walk(abs));
+        } else if (e.isFile()) {
+            if (!isMarkdown(abs)) continue;
+            if (e.name.includes('sync-conflict')) continue;
+            files.push(abs);
+        }
+    }
+    return files;
+}
+
+export async function reindexAll(): Promise<number> {
+    try {
+        const absPaths = await walk(CONFIG.vaultRoot);
+        const docs = [] as any[];
+        for (const abs of absPaths) {
+            const rel = path.relative(CONFIG.vaultRoot, abs).split(path.sep).join('/');
+            const buf = await fs.readFile(abs, 'utf8');
+            const parsed = matter(buf);
+            const stat = await fs.stat(abs);
+            docs.push(toSearchDoc({ path: rel, frontmatter: parsed.data ?? {}, content: parsed.content, mtime: stat.mtimeMs }));
+        }
+        if (docs.length) await index.addDocuments(docs);
+        return docs.length;
+    } catch (err: any) {
+        if (err?.code === 'ENOENT') return 0;
+        throw err;
+    }
+}
